Fix #include_next recursion if there's no next
If a system header uses #include_next to include a header for which there is
no next file, it would include the same header again due to the includes list
including all of the system includes, and then the recurse for #include_next
appending the system headers again.

This commit fixes that by no longer re-appending the system headers and making
sure duplicates are removed when they are included in the first place. Now a
missing include error will happen instead.