Stop including system headers over local headers
The directory of the current file is appended to the end of the list of header
directories. If a system header (or any header in the include paths) is named
the same as a local header included with "", the system header will be
included instead of the local. By prepending the header insted, the local one
will be found first.

Additionally headers included with <> should only use the system headers and
include paths, so it should not include local files.