Introduce a new expression interface with a few basic types.
The change defines a new Expr interface that is intended to simplify
access to the C AST exposed by the package. The new interface has a
single placeholder method that does nothing. Future commits will define
the interface in more details.

Also, introduce a few basic expression types and methods to convert
*Expression types to the new interface type. For now those methods
mostly panic with a TODO error. Future changes will address those.

Defined types are not complete by any means. For example, Ident
needs a reference to a scope where it was defined, the type, etc.

Signed-off-by: Denys Smirnov <denis.smirnov.91@gmail.com>